Refactor of convolution class
- Created separated wrapper for HD to make it easier work on VD
- Removed unused variables and fixed few mistakes
- Added convolution option with FFT. Faster (nlog(n)) but have a weird baseline due to undershoot
